Fix kqueue on platforms where C's long is i32
From (my own) commit to mio: 87d72418338a9250f70cb4a43c8bea10371f8256.

+++ src/sys/unix/kqueue.rs
@@ -203,7 +203,11 @@ impl Selector {
 fn timespec_from_duration(duration: Duration) -> libc::timespec {
     libc::timespec {
         tv_sec: min(duration.as_secs(), libc::time_t::max_value() as u64) as libc::time_t,
-        tv_nsec: libc::c_long::from(duration.subsec_nanos()),
+        // `Duration::subsec_nanos` is guaranteed to be less than one
+        // billion (the number of nanoseconds in a second), making the
+        // cast to i32 safe. The cast itself is needed for platforms
+        // where C's long is only 32 bits.
+        tv_nsec: libc::c_long::from(duration.subsec_nanos() as i32),
     }
 }
